+++ dbt/include/databricks/macros/adapters.sql
@@ -1,19 +1,11 @@
-{% macro dbt_databricks_file_format_clause() %}
+{% macro databricks__file_format_clause() %}
   {%- set file_format = config.get('file_format', default='delta') -%}
   {%- if file_format is not none %}
     using {{ file_format }}
   {%- endif %}
 {%- endmacro -%}
 
-{% macro dbt_databricks_location_clause() %}
-  {%- set location_root = config.get('location_root', validator=validation.any[basestring]) -%}
-  {%- set identifier = model['alias'] -%}
-  {%- if location_root is not none %}
-    location '{{ location_root }}/{{ identifier }}'
-  {%- endif %}
-{%- endmacro -%}
-
-{% macro dbt_databricks_options_clause() -%}
+{% macro databricks__options_clause() -%}
   {%- set options = config.get('options') -%}
   {%- if config.get('file_format', default='delta') == 'hudi' -%}
     {%- set unique_key = config.get('unique_key') -%}
@@ -35,51 +27,12 @@
   {%- endif %}
 {%- endmacro -%}
 
-{% macro dbt_databricks_comment_clause() %}
-  {%- set raw_persist_docs = config.get('persist_docs', {}) -%}
 
-  {%- if raw_persist_docs is mapping -%}
-    {%- set raw_relation = raw_persist_docs.get('relation', false) -%}
-      {%- if raw_relation -%}
-      comment '{{ model.description | replace("'", "\\'") }}'
-      {% endif %}
-  {%- elif raw_persist_docs -%}
-    {{ exceptions.raise_compiler_error("Invalid value provided for 'persist_docs'. Expected dict but got value: " ~ raw_persist_docs) }}
-  {% endif %}
+{% macro tblproperties_clause() -%}
+  {{ return(adapter.dispatch('tblproperties_clause', 'dbt')()) }}
 {%- endmacro -%}
 
-{% macro dbt_databricks_partition_cols(label, required=false) %}
-  {%- set cols = config.get('partition_by', validator=validation.any[list, basestring]) -%}
-  {%- if cols is not none %}
-    {%- if cols is string -%}
-      {%- set cols = [cols] -%}
-    {%- endif -%}
-    {{ label }} (
-    {%- for item in cols -%}
-      {{ item }}
-      {%- if not loop.last -%},{%- endif -%}
-    {%- endfor -%}
-    )
-  {%- endif %}
-{%- endmacro -%}
-
-{% macro dbt_databricks_clustered_cols(label, required=false) %}
-  {%- set cols = config.get('clustered_by', validator=validation.any[list, basestring]) -%}
-  {%- set buckets = config.get('buckets', validator=validation.any[int]) -%}
-  {%- if (cols is not none) and (buckets is not none) %}
-    {%- if cols is string -%}
-      {%- set cols = [cols] -%}
-    {%- endif -%}
-    {{ label }} (
-    {%- for item in cols -%}
-      {{ item }}
-      {%- if not loop.last -%},{%- endif -%}
-    {%- endfor -%}
-    ) into {{ buckets }} buckets
-  {%- endif %}
-{%- endmacro -%}
-
-{% macro dbt_databricks_tblproperties_clause() -%}
+{% macro databricks__tblproperties_clause() -%}
   {%- set tblproperties = config.get('tblproperties') -%}
   {%- if tblproperties is not none %}
     tblproperties (
@@ -90,28 +43,23 @@
   {%- endif %}
 {%- endmacro -%}
 
-{#-- We can't use temporary tables with `create ... as ()` syntax #}
-{% macro dbt_databricks_create_temporary_view(relation, sql) -%}
-  create temporary view {{ relation.include(schema=false) }} as
-    {{ sql }}
-{% endmacro %}
 
 {% macro databricks__create_table_as(temporary, relation, sql) -%}
   {% if temporary -%}
-    {{ dbt_databricks_create_temporary_view(relation, sql) }}
+    {{ create_temporary_view(relation, sql) }}
   {%- else -%}
     {% if config.get('file_format', default='delta') == 'delta' %}
       create or replace table {{ relation }}
     {% else %}
       create table {{ relation }}
     {% endif %}
-    {{ dbt_databricks_file_format_clause() }}
-    {{ dbt_databricks_options_clause() }}
-    {{ dbt_databricks_partition_cols(label="partitioned by") }}
-    {{ dbt_databricks_clustered_cols(label="clustered by") }}
-    {{ dbt_databricks_location_clause() }}
-    {{ dbt_databricks_comment_clause() }}
-    {{ dbt_databricks_tblproperties_clause() }}
+    {{ file_format_clause() }}
+    {{ options_clause() }}
+    {{ partition_cols(label="partitioned by") }}
+    {{ clustered_cols(label="clustered by") }}
+    {{ location_clause() }}
+    {{ comment_clause() }}
+    {{ tblproperties_clause() }}
     as
       {{ sql }}
   {%- endif %}
@@ -119,8 +67,8 @@
 
 {% macro databricks__create_view_as(relation, sql) -%}
   create or replace view {{ relation }}
-  {{ dbt_databricks_comment_clause() }}
-  {{ dbt_databricks_tblproperties_clause() }}
+  {{ comment_clause() }}
+  {{ tblproperties_clause() }}
   as
     {{ sql }}
 {% endmacro %}
@@ -131,11 +79,67 @@
       {% set comment = column_dict[column_name]['description'] %}
       {% set escaped_comment = comment | replace('\'', '\\\'') %}
       {% set comment_query %}
-        alter table {{ relation }} change column 
+        alter table {{ relation }} change column
             {{ adapter.quote(column_name) if column_dict[column_name]['quote'] else column_name }}
             comment '{{ escaped_comment }}';
       {% endset %}
       {% do run_query(comment_query) %}
     {% endfor %}
   {% endif %}
+{% endmacro %}
+
+{# Persist table-level and column-level constraints. #}
+{% macro persist_constraints(relation, model) %}
+  {{ return(adapter.dispatch('persist_constraints', 'dbt')(relation, model)) }}
+{% endmacro %}
+
+{% macro databricks__persist_constraints(relation, model) %}
+  {% if config.get('persist_constraints', False) and config.get('file_format', 'delta') == 'delta' %}
+    {% do alter_table_add_constraints(relation, model.meta.constraints) %}
+    {% do alter_column_set_constraints(relation, model.columns) %}
+  {% endif %}
+{% endmacro %}
+
+{% macro alter_table_add_constraints(relation, constraints) %}
+  {{ return(adapter.dispatch('alter_table_add_constraints', 'dbt')(relation, constraints)) }}
+{% endmacro %}
+
+{% macro databricks__alter_table_add_constraints(relation, constraints) %}
+  {% if constraints is sequence %}
+    {% for constraint in constraints %}
+      {% set name = constraint['name'] %}
+      {% if not name %}
+        {{ exceptions.raise_compiler_error('Invalid check constraint name: ' ~ name) }}
+      {% endif %}
+      {% set condition = constraint['condition'] %}
+      {% if not condition %}
+        {{ exceptions.raise_compiler_error('Invalid check constraint condition: ' ~ condition) }}
+      {% endif %}
+      {# Skip if the update is incremental. #}
+      {% if not is_incremental() %}
+        {% call statement() %}
+          alter table {{ relation }} add constraint {{ name }} check ({{ condition }});
+        {% endcall %}
+      {% endif %}
+    {% endfor %}
+  {% endif %}
+{% endmacro %}
+
+{% macro alter_column_set_constraints(relation, column_dict) %}
+  {{ return(adapter.dispatch('alter_column_set_constraints', 'dbt')(relation, column_dict)) }}
+{% endmacro %}
+
+{% macro databricks__alter_column_set_constraints(relation, column_dict) %}
+  {% for column_name in column_dict %}
+    {% set constraint = column_dict[column_name]['meta']['constraint'] %}
+    {% if constraint %}
+      {% if constraint != 'not_null' %}
+        {{ exceptions.raise_compiler_error('Invalid constraint for column ' ~ column_name ~ '. Only `not_null` is supported.') }}
+      {% endif %}
+      {% set quoted_name = adapter.quote(column_name) if column_dict[column_name]['quote'] else column_name %}
+      {% call statement() %}
+        alter table {{ relation }} change column {{ quoted_name }} set not null
+      {% endcall %}
+    {% endif %}
+  {% endfor %}
 {% endmacro %}