+++ dbt/include/databricks/macros/materializations/snapshot.sql
@@ -96,6 +96,8 @@
 
   {% do persist_docs(target_relation, model) %}
 
+  {% do persist_constraints(target_relation, model) %}
+
   {{ run_hooks(post_hooks, inside_transaction=True) }}
 
   {{ adapter.commit() }}