+++ tests/integration/persist_constraints/models/invalid_column_constraint.sql
@@ -1,0 +1,4 @@
+{{config(materialized='table')}}
+
+select * from {{ ref('seed') }}
+