+++ tests/integration/seed_column_types/test_seed_column_types.py
@@ -1,13 +1,11 @@
-from cProfile import run
 from tests.integration.base import DBTIntegrationTest, use_profile
-import dbt.exceptions
 
 
 class TestSeedColumnTypeCast(DBTIntegrationTest):
     @property
     def schema(self):
         return "seed_column_types"
-        
+
     @property
     def models(self):
         return "models"
@@ -15,8 +13,8 @@ class TestSeedColumnTypeCast(DBTIntegrationTest):
     @property
     def project_config(self):
         return {
-            'seeds': {
-                'quote_columns': False,
+            "seeds": {
+                "quote_columns": False,
             },
         }
 
@@ -24,6 +22,14 @@ class TestSeedColumnTypeCast(DBTIntegrationTest):
     def test_seed_column_types_databricks_cluster(self):
         self.run_dbt(["seed"])
 
+    @use_profile("databricks_uc_cluster")
+    def test_seed_column_types_databricks_uc_cluster(self):
+        self.run_dbt(["seed"])
+
     @use_profile("databricks_sql_endpoint")
     def test_seed_column_types_databricks_sql_endpoint(self):
+        self.run_dbt(["seed"])
+
+    @use_profile("databricks_uc_sql_endpoint")
+    def test_seed_column_types_databricks_uc_sql_endpoint(self):
         self.run_dbt(["seed"])